message data for a given time period completed
