+++ server/slackMethods.js
@@ -31,6 +31,8 @@ export async function getChannelMessages() {
 		return await client.conversations.history({
 			token: process.env.TOKEN,
 			channel: process.env.CHANNEL_ID,
+      oldest: 1676377207.351409,
+      latest:1676897346.238299,
 		});
 	} catch (error) {
 		console.error(error);
